changed where some of the ts files are locationed. refactored the workout page with minute times
